render: store sprite/glyph vertex positions as float
The GL ES 1.1 backend specifies the vertex position array as GL_FLOAT
(glVertexPointer type 0x1406, verified in FUN_0001342c), and the primitive
path (neColorVertex) already uses float positions. But neTexVertex (textured
sprites) and neGlyphVertex (text) declared the position as int32_t while
being fed through the same GL_FLOAT pointer, so an integer coordinate such as
640 was reinterpreted as an IEEE float (~9e-43, a denormal near zero),
collapsing every textured quad and glyph to the origin.

Change both position fields to float; the aggregate initializers and the
integer pen expressions auto-convert. Stride stays 16 bytes, so the UV
(GL_SHORT) and colour (GL_UNSIGNED_BYTE) offsets are unchanged.

Necessary but not sufficient for visible sprites: the stretch-sprite
parameter chain (neTextureForiOS::draw -> drawSprite ->
drawAepOtSpriteStretch -> drawAepSpriteClipped) still forwards scrambled
fields and is being re-reconstructed against Ghidra separately.

+++ Project/System/src/Render/neRenderer.cpp
@@ -379,12 +379,15 @@ void neDeleteTexture(neRenderer *r, int name) {
 // Textured sprite blit.
 // ---------------------------------------------------------------------------
 
-// One textured vertex: 16.16 fixed-point position, GL_SHORT-normalized UV,
-// premult RGBA8. Stride 16; position size 2 (+0x00), UV (+0x08), colour
-// (+0x0c). Ghidra: built on the stack (local_a4 block) by FUN_00015fb8.
+// One textured vertex: GL_FLOAT position, GL_SHORT-normalized UV, premult RGBA8.
+// Stride 16; position size 2 (+0x00), UV (+0x08), colour (+0x0c). Ghidra: built
+// on the stack (local_a4 block) by FUN_00015fb8; the backend specifies the
+// position array as GL_FLOAT (0x1406), so the corners must be stored as floats —
+// storing ints here would be reinterpreted as near-zero denormals and collapse
+// every quad to the origin.
 struct neTexVertex {
-    int32_t x;
-    int32_t y;
+    float x;
+    float y;
     int16_t u;
     int16_t v;
     uint32_t rgba;
